Drop Docker Compose support, add local Gradle caching via Docker volume
- drop Docker Compose support due to issues with the new Dockerfile
  and scripts setup, and how the JAR file is transported to the
  main container,
- store Gradle cache in a Docker volume to speed up the rebuilds of
  the application,
- clean up the scripts,
- update readme.

+++ bin/teardown.sh
@@ -1,52 +1,78 @@
 #!/usr/bin/env sh
 
 MAIN_IMAGE='renameme'
-NETWORK="${MAIN_IMAGE}-network"
-DATABASE="${MAIN_IMAGE}-database"
-CACHE_VOLUME='gradle-cache'
-TEMP_DIRECTORIES='.gradle build'
+MAIN_CONTAINER="$MAIN_IMAGE"
 
-NOFORMAT='\033[0m' PURPLE='\033[0;35m'
+BUILD_IMAGE="${MAIN_IMAGE}-gradle-build"
+BUILD_CONTAINER="$BUILD_IMAGE"
+
+DATABASE_CONTAINER="${MAIN_IMAGE}-database"
+
+NETWORK="${MAIN_IMAGE}-network"
+CACHE_VOLUME='gradle-build-cache'
+TEMP_DIRECTORIES='.gradle build tmp'
+
+log() {
+    COLOUR_RESET='\033[0m'
+    PURPLE_BOLD='\033[1;35m'
+    echo "${PURPLE_BOLD}==> ${1} <==${COLOUR_RESET}"
+}
+
+error() {
+    COLOUR_RESET='\033[0m'
+    RED_BOLD='\033[1;31m'
+    echo "${RED_BOLD}==> ${1} <==${COLOUR_RESET}"
+    exit 1
+}
 
 while [ "$#" -gt 0 ]; do
     case $1 in
-    --exclude-db)
-        _exclude_db='true'
-        ;;
     --include-cache)
         _include_cache='true'
         ;;
+    --include-db)
+        _include_db='true'
+        ;;
     *)
-        echo "Unknown option: ${1}"
-        exit 1
+        error "Unknown option: '${1}'"
         ;;
     esac
     shift
 done
 
-echo "${PURPLE}Stopping '${MAIN_IMAGE}' container${NOFORMAT}"
-docker container stop "$MAIN_IMAGE"
-echo "${PURPLE}Removing '${MAIN_IMAGE}' container${NOFORMAT}"
-docker container rm --volumes "$MAIN_IMAGE"
-echo "${PURPLE}Removing '${MAIN_IMAGE}' image${NOFORMAT}"
+log "Stopping '${BUILD_CONTAINER}' container"
+docker container stop "$BUILD_CONTAINER"
+log "Removing '${BUILD_CONTAINER}' container"
+docker container rm "$BUILD_CONTAINER"
+log "Removing '${BUILD_IMAGE}' image"
+docker image rm "$BUILD_IMAGE"
+
+log "Stopping '${MAIN_CONTAINER}' container"
+docker container stop "$MAIN_CONTAINER"
+log "Removing '${MAIN_CONTAINER}' container"
+docker container rm --volumes "$MAIN_CONTAINER"
+log "Removing '${MAIN_IMAGE}' image"
 docker image rm "$MAIN_IMAGE"
 
-echo "${PURPLE}Stopping '${DATABASE}' container${NOFORMAT}"
-docker container stop "$DATABASE"
-if [ 'true' != "$_exclude_db" ]; then
-    echo "${PURPLE}Removing '${DATABASE}' container${NOFORMAT}"
-    docker container rm --volumes "$DATABASE"
+if [ 'true' = "$_include_db" ]; then
+    log "Stopping '${DATABASE_CONTAINER}' container"
+    docker container stop "$DATABASE_CONTAINER"
+    log "Removing '${DATABASE_CONTAINER}' container"
+    docker container rm --volumes "$DATABASE_CONTAINER"
 fi
 
-echo "${PURPLE}Removing '${NETWORK}' network${NOFORMAT}"
-docker network rm "$NETWORK"
+# Since nothing exists within the network at this point, remove it
+if [ 'true' = "$_include_db" ]; then
+    log "Removing '${NETWORK}' network"
+    docker network rm "$NETWORK"
+fi
 
 if [ 'true' = "$_include_cache" ]; then
-    echo "${PURPLE}Removing '${CACHE_VOLUME}' volume${NOFORMAT}"
+    log "Removing '${CACHE_VOLUME}' volume"
     docker volume rm "$CACHE_VOLUME"
 fi
 
 for directory in $TEMP_DIRECTORIES; do
-    echo "${PURPLE}Removing '${directory}'${NOFORMAT}"
+    log "Removing '${directory}' temp directory"
     rm -rf "$directory"
 done
